requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.4.0
astunparse==1.6.3
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.3
bidict==0.23.1
blinker==1.8.2
boto3==1.34.122
botocore==1.34.122
bson==0.5.10
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.4.5
contourpy==1.2.1
cryptography==42.0.8
cycler==0.12.1
dataclasses-json==0.6.6
deepgram-sdk==3.2.7
distlib==0.3.8
distro==1.9.0
dnspython==2.6.1
email_validator==2.1.1
et-xmlfile==1.1.0
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.15.4
firebase-admin==6.5.0
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-RESTful==0.3.10
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
gast==0.6.0
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.132.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-generativeai==0.6.0
google-pasta==0.2.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.1
gql==3.5.0
graphql-core==3.2.3
grpcio==1.64.1
grpcio-status==1.62.2
gunicorn==20.1.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
keras==3.4.1
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.1
mdurl==0.1.2
ml-dtypes==0.4.0
MouseInfo==0.1.3
msgpack==1.0.8
multidict==6.0.5
mypy-extensions==1.0.0
namex==0.0.8
newrelic==9.11.0
nltk==3.8.1
numpy==1.26.4
openai==1.37.2
opencv-python==4.10.0.84
openpyxl==3.1.5
opt-einsum==3.3.0
optree==0.12.1
orjson==3.10.4
packaging==24.0
pandas==2.2.2
pillow==10.4.0
pipenv==2024.0.1
platformdirs==4.2.2
pluggy==1.5.0
proto-plus==1.23.0
protobuf==4.25.3
psutil==6.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAutoGUI==0.9.54
pycaw==20240210
pycparser==2.22
pydantic==2.8.2
pydantic-extra-types==2.8.0
pydantic-settings==2.3.2
pydantic_core==2.20.1
PyGetWindow==0.0.9
Pygments==2.18.0
pyjokes==0.7.2
PyJWT==2.8.0
pymongo==4.7.3
PyMsgBox==1.0.9
pyparsing==3.1.2
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==0.1.30
pytest==8.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-multipart==0.0.9
python-socketio==5.11.2
pytweening==1.2.0
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
requests-aws4auth==1.3.1
requests-toolbelt==1.0.0
rich==13.7.1
rsa==4.9
s3transfer==0.10.1
schedule==1.2.2
setuptools==70.3.0
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
slack_sdk==3.31.0
sniffio==1.3.1
starlette==0.37.2
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
termcolor==2.4.0
tqdm==4.66.4
typer==0.12.3
typing==3.7.4.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.30.1
verboselogs==1.7
virtualenv==20.26.3
watchfiles==0.22.0
websockets==12.0
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
